django_backend/requirements.txt
Django==5.1.7
djangorestframework==3.15.2
djoser==2.3.1
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
Pillow==11.3.0
django-filter==24.3